Move each thunk action creators
aiming to the readable code
