+++ frontend/src/components/Home/GuestAction.tsx
@@ -5,7 +5,8 @@ import { Button, List, ListItem } from '@material-ui/core';
 import { Menu as MenuIcon } from '@material-ui/icons';
 import PopoverControl from '../../templates/PopoverControl';
 import { useAppDispatch } from '../../store/hooks';
-import { createUser, signInWithEmail } from '../../store/slices/authSlice';
+import { createUser } from 'store/thunks/createUser';
+import { signInWithEmail } from 'store/thunks/signInWithEmail';
 
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
