+++ frontend/src/pages/auth/EmailVerification.tsx
@@ -7,10 +7,8 @@ import Header from '../../layouts/Header';
 import Footer from '../../layouts/Footer';
 import { APP_NAME } from '../../config/app';
 import { useAppDispatch, useAppSelector } from '../../store/hooks';
-import {
-  deleteSentEmailState,
-  sendEmailVerificationLink,
-} from '../../store/slices/authSlice';
+import { deleteSentEmailState } from 'store/slices/authSlice';
+import { sendEmailVerificationLink } from 'store/thunks/sendEmailVerificationLink';
 import { isSentEmail, isVerified } from '../../utils/auth';
 
 const useStyles = makeStyles((theme: Theme) =>
