+++ frontend/src/store/slices/authSlice.ts
@@ -1,382 +1,17 @@
-import { createAsyncThunk, createSlice, PayloadAction } from '@reduxjs/toolkit';
-import axios, { AxiosError, AxiosResponse } from 'axios';
-import {
-  API_HOST,
-  FORGOT_PASSWORD_PATH,
-  AUTH_USER_PATH,
-  GET_CSRF_TOKEN_PATH,
-  RESET_PASSWORD_PATH,
-  SIGNIN_PATH,
-  SIGNOUT_PATH,
-  SIGNUP_PATH,
-  UPDATE_PASSWORD_PATH,
-  UPDATE_USER_INFO_PATH,
-  VERIFICATION_NOTIFICATION_PATH,
-} from '../../config/api';
-import { User } from '../../models/User';
+import { createSlice, PayloadAction } from '@reduxjs/toolkit';
+
+import { User } from 'models/User';
 import { FlashNotificationProps } from 'layouts/FlashNotification';
-
-// `store`の利用不可、それを利用した関数も同様 (以下のエラー発生)
-// TypeError: Cannot read property 'reducer' of undefined
-const authApiClient = axios.create({
-  baseURL: API_HOST,
-  withCredentials: true,
-});
-
-type RejectWithValueType = {
-  error: {
-    message?: string;
-    data: any;
-  };
-};
-
-export type SignUpRequest = {
-  name: string;
-  email: string;
-  password: string;
-  password_confirmation: string;
-};
-
-export type SignUpResponse = {
-  user: User;
-};
-
-export const createUser = createAsyncThunk<
-  SignUpResponse,
-  SignUpRequest,
-  { rejectValue: RejectWithValueType }
->('auth/createUser', async (payload, thunkApi) => {
-  const { email, password, password_confirmation } = payload;
-  try {
-    await authApiClient.get(GET_CSRF_TOKEN_PATH);
-    const response = await authApiClient.post(
-      SIGNUP_PATH,
-      {
-        name: email,
-        email,
-        password,
-        password_confirmation,
-      },
-      { validateStatus: (status) => status === 201 }
-    );
-    return response?.data as SignUpResponse;
-    // 正常時はステータスコード`201`, `response.data`なし
-  } catch (e) {
-    // ステータスコード 2xx 以外を`catch`
-    const error: AxiosError = e; // cast the error for access
-    if (error.response?.status === 422) {
-      return thunkApi.rejectWithValue({
-        error: {
-          // 他のバリデーションはフロントエンドで実施
-          message: 'このメールアドレスは既に使用されています',
-          data: error.response.data,
-        },
-      });
-    } // `authSlice`の`extraReducers`で`rejected`を呼び出す
-    return thunkApi.rejectWithValue({
-      error: {
-        message: 'システムエラーが発生しました',
-        data: error?.response?.data,
-      },
-    });
-  }
-});
-
-export type FetchAuthUserResponse = {
-  user: User;
-};
-
-export const fetchAuthUser = createAsyncThunk<
-  FetchAuthUserResponse,
-  void,
-  { rejectValue: RejectWithValueType }
->('auth/fetchAuthUser', async (_, thunkApi) => {
-  try {
-    const response = await authApiClient.get(AUTH_USER_PATH);
-    return response?.data as FetchAuthUserResponse;
-  } catch (e) {
-    const error: AxiosError = e;
-    return thunkApi.rejectWithValue({
-      error: { data: error?.response?.data },
-    });
-  }
-});
-
-export const sendEmailVerificationLink = createAsyncThunk<
-  AxiosResponse<any>['status'],
-  void,
-  { rejectValue: RejectWithValueType }
->('auth/sendVerificationLink', async (_, thunkApi) => {
-  try {
-    // 正常時は`202`(認証済みの場合`204`)
-    const response = await authApiClient.post(VERIFICATION_NOTIFICATION_PATH);
-    return response.status;
-  } catch (e) {
-    const error: AxiosError = e;
-    if (error.response && [401, 419].includes(error.response.status)) {
-      thunkApi.dispatch(signOut());
-      thunkApi.dispatch(
-        setFlash({ type: 'error', message: 'ログインしてください' })
-      );
-    }
-    return thunkApi.rejectWithValue({
-      error: {
-        message: 'システムエラーが発生しました',
-        data: error?.response?.data,
-      },
-    });
-  }
-});
-
-export type SignInResponse = {
-  user: User;
-  verified: true | undefined;
-};
-
-export type SignInRequest = {
-  email: string;
-  password: string;
-  remember: string | undefined;
-};
-
-export const signInWithEmail = createAsyncThunk<
-  SignInResponse,
-  SignInRequest,
-  { rejectValue: RejectWithValueType }
->('auth/signInWithEmail', async (payload, thunkApi) => {
-  const { email, password, remember } = payload;
-  try {
-    await authApiClient.get(GET_CSRF_TOKEN_PATH);
-    const response = await authApiClient.post(SIGNIN_PATH, {
-      email,
-      password,
-      remember,
-    });
-    return response?.data;
-  } catch (e) {
-    const error: AxiosError = e;
-    if (error.response?.status === 422) {
-      return thunkApi.rejectWithValue({
-        error: {
-          message: 'メールアドレスまたはパスワードが間違っています',
-          data: error.response.data,
-        },
-      });
-    } else if (error.response?.status === 429) {
-      return thunkApi.rejectWithValue({
-        error: {
-          message:
-            '所定回数を超えて誤った入力が行われたため、アクセスを制限しております',
-          data: error.response.data,
-        },
-      });
-    } // 認証用メールから遷移して、認証リンクが無効だった場合
-    else if (error.response?.status === 403) {
-      thunkApi.dispatch(fetchAuthUser());
-      thunkApi.dispatch(
-        setFlash({ type: 'warning', message: '認証に失敗しました' })
-      );
-    }
-    return thunkApi.rejectWithValue({
-      error: {
-        message: 'システムエラーが発生しました',
-        data: error?.response?.data,
-      },
-    });
-  }
-});
-
-export const updateProfile = createAsyncThunk<
-  { username: string; email: string },
-  { username: string; email: string },
-  { rejectValue: RejectWithValueType }
->('auth/updateProfile', async (payload, thunkApi) => {
-  const { username, email } = payload;
-  try {
-    await authApiClient.put(UPDATE_USER_INFO_PATH, {
-      name: username,
-      email,
-    });
-    return { username, email };
-  } catch (e) {
-    const error: AxiosError = e;
-    if (error.response && [401, 419].includes(error.response.status)) {
-      thunkApi.dispatch(signOut());
-      thunkApi.dispatch(
-        setFlash({ type: 'error', message: 'ログインしてください' })
-      );
-    }
-    if (error.response?.status === 422) {
-      return thunkApi.rejectWithValue({
-        error: {
-          message: 'このメールアドレスは既に使用されています',
-          data: error.response.data,
-        },
-      });
-    }
-    return thunkApi.rejectWithValue({
-      error: {
-        message: 'システムエラーが発生しました',
-        data: error?.response?.data,
-      },
-    });
-  }
-});
-
-export const updatePassword = createAsyncThunk<
-  void,
-  { current_password: string; password: string; password_confirmation: string },
-  { rejectValue: RejectWithValueType }
->('auth/updatePassword', async (payload, thunkApi) => {
-  const { current_password, password, password_confirmation } = payload;
-  try {
-    await authApiClient.put(UPDATE_PASSWORD_PATH, {
-      current_password,
-      password,
-      password_confirmation,
-    });
-  } catch (e) {
-    const error: AxiosError = e;
-    if (error.response && [401, 419].includes(error.response.status)) {
-      thunkApi.dispatch(signOut());
-      thunkApi.dispatch(
-        setFlash({ type: 'error', message: 'ログインしてください' })
-      );
-    }
-    if (error.response?.status === 422) {
-      return thunkApi.rejectWithValue({
-        error: {
-          message: 'パスワードが間違っています',
-          data: error.response.data,
-        },
-      });
-    }
-    return thunkApi.rejectWithValue({
-      error: {
-        message: 'システムエラーが発生しました',
-        data: error?.response?.data,
-      },
-    });
-  }
-});
-
-export const forgotPassword = createAsyncThunk<
-  string,
-  { email: string },
-  { rejectValue: RejectWithValueType }
->('auth/forgotPassword', async (payload, thunkApi) => {
-  const { email } = payload;
-  try {
-    // 正常時は`200`バリデーションエラー時は`422`
-    await authApiClient.get(GET_CSRF_TOKEN_PATH);
-    const response = await authApiClient.post(FORGOT_PASSWORD_PATH, {
-      email,
-    });
-    return response.data?.message;
-  } catch (e) {
-    const error: AxiosError = e;
-    if (error.response?.status === 422) {
-      return thunkApi.rejectWithValue({
-        error: {
-          message: '指定されたメールアドレスは存在しません',
-          data: error.response.data,
-        },
-      });
-    }
-    return thunkApi.rejectWithValue({
-      error: {
-        message: 'システムエラーが発生しました',
-        data: error?.response?.data,
-      },
-    });
-  }
-});
-
-export const resetPassword = createAsyncThunk<
-  void,
-  {
-    email: string;
-    password: string;
-    password_confirmation: string;
-    token: string;
-  },
-  { rejectValue: RejectWithValueType }
->('auth/resetPassword', async (payload, thunkApi) => {
-  const { email, password, password_confirmation, token } = payload;
-  try {
-    // 正常時は`200`バリデーションエラー時は`422`
-    await authApiClient.get(GET_CSRF_TOKEN_PATH);
-    await authApiClient.post(RESET_PASSWORD_PATH, {
-      email,
-      password,
-      password_confirmation,
-      token,
-    });
-  } catch (e) {
-    const error: AxiosError = e;
-    if (error.response?.status === 422) {
-      return thunkApi.rejectWithValue({
-        error: {
-          message: '認証に失敗しました\n再度お試しください',
-          data: error.response.data,
-        },
-      });
-    }
-    return thunkApi.rejectWithValue({
-      error: {
-        message: 'システムエラーが発生しました',
-        data: error?.response?.data,
-      },
-    });
-  }
-});
-
-export const putSignOut = createAsyncThunk<
-  void,
-  void,
-  { rejectValue: RejectWithValueType }
->('auth/putSignOut', async (_, thunkApi) => {
-  try {
-    // status(response): ログイン状態によらず`204` 認証切れなら`419`
-    await authApiClient.post(SIGNOUT_PATH);
-  } catch (e) {
-    const error: AxiosError = e;
-    if (error.response && [401, 419].includes(error.response.status)) {
-      thunkApi.dispatch(signOut());
-    }
-    return thunkApi.rejectWithValue({
-      error: {
-        message: 'システムエラーが発生しました',
-        data: error?.response?.data,
-      },
-    });
-  }
-});
-
-export const deleteAccount = createAsyncThunk<
-  any,
-  void,
-  { rejectValue: RejectWithValueType }
->('auth/deleteAccount', async (_, thunkApi) => {
-  try {
-    await authApiClient.delete(AUTH_USER_PATH);
-  } catch (e) {
-    const error: AxiosError = e;
-    if (error.response && [401, 419].includes(error.response.status)) {
-      thunkApi.dispatch(signOut());
-      thunkApi.dispatch(
-        setFlash({ type: 'error', message: 'ログインしてください' })
-      );
-    }
-    setFlash({ type: 'error', message: 'システムエラーが発生しました' });
-    return thunkApi.rejectWithValue({
-      error: {
-        message: 'システムエラーが発生しました',
-        data: error?.response?.data,
-      },
-    });
-  }
-});
+import { createUser } from 'store/thunks/createUser';
+import { fetchAuthUser } from 'store/thunks/fetchAuthUser';
+import { sendEmailVerificationLink } from 'store/thunks/sendEmailVerificationLink';
+import { signInWithEmail } from 'store/thunks/signInWithEmail';
+import { updateProfile } from 'store/thunks/updateProfile';
+import { updatePassword } from 'store/thunks/updatePassword';
+import { forgotPassword } from 'store/thunks/forgotPassoword';
+import { resetPassword } from 'store/thunks/resetPassword';
+import { putSignOut } from 'store/thunks/putSignOut';
+import { deleteAccount } from 'store/thunks/deleteAccount';
 
 type AuthState = {
   user: User | null;
