+++ frontend/src/store/thunks/index.ts
@@ -1,0 +1,8 @@
+// `createAsyncThunk` returns a standard Redux thunk action creator.
+
+export type RejectWithValueType = {
+  error: {
+    message?: string;
+    data: any;
+  };
+};
