+++ frontend/src/store/thunks/resetPassword.ts
@@ -1,0 +1,47 @@
+import { createAsyncThunk } from '@reduxjs/toolkit';
+import { AxiosError } from 'axios';
+
+import { GET_CSRF_TOKEN_PATH, RESET_PASSWORD_PATH } from 'config/api';
+import { authApiClient } from './utils/api';
+import { RejectWithValueType } from '.';
+
+export const resetPassword = createAsyncThunk<
+  void,
+  {
+    email: string;
+    password: string;
+    password_confirmation: string;
+    token: string;
+  },
+  { rejectValue: RejectWithValueType }
+>('auth/resetPassword', async (payload, thunkApi) => {
+  const { email, password, password_confirmation, token } = payload;
+  try {
+    // 正常時は`200`バリデーションエラー時は`422`
+    await authApiClient.get(GET_CSRF_TOKEN_PATH);
+    await authApiClient.post(RESET_PASSWORD_PATH, {
+      email,
+      password,
+      password_confirmation,
+      token,
+    });
+  } catch (e) {
+    const error: AxiosError = e;
+    if (error.response?.status === 422) {
+      return thunkApi.rejectWithValue({
+        error: {
+          message: '認証に失敗しました\n再度お試しください',
+          data: error.response.data,
+        },
+      });
+    }
+    return thunkApi.rejectWithValue({
+      error: {
+        message: 'システムエラーが発生しました',
+        data: error?.response?.data,
+      },
+    });
+  }
+});
+
+export default resetPassword;
